+++ packages/atomic/src/components/atomic-result-template/atomic-result-template.tsx
@@ -1,13 +1,7 @@
-import {Component, Element, Prop, Method, State, h} from '@stencil/core';
-import {
-  ResultTemplate,
-  ResultTemplateCondition,
-  ResultTemplatesHelpers,
-} from '@coveo/headless';
+import {Component, Element, Prop, Method, State} from '@stencil/core';
+import {ResultTemplate, ResultTemplateCondition} from '@coveo/headless';
 import {MapProp} from '../../utils/props-utils';
-import {Bindings, InitializeBindings} from '../../utils/initialization-utils';
-
-export type TemplateContent = DocumentFragment;
+import {ResultTemplateCommon} from '../result-templates/result-template-common';
 
 /**
  * The `atomic-result-template` component determines the format of the query results, depending on the conditions that are defined for each template. A `template` element must be the child of an `atomic-result-template`, and an `atomic-result-list` must be the parent of each `atomic-result-template`.
@@ -19,12 +13,9 @@ export type TemplateContent = DocumentFragment;
   shadow: true,
 })
 export class AtomicResultTemplate {
-  @InitializeBindings() public bindings!: Bindings;
   @State() public error!: Error;
 
-  private matchConditions: ResultTemplateCondition[] = [];
-
-  @Element() private host!: HTMLDivElement;
+  @Element() public host!: HTMLDivElement;
 
   /**
    * A function that must return true on results for the result template to apply.
@@ -49,88 +40,35 @@ export class AtomicResultTemplate {
   @MapProp({splitValues: true}) public mustNotMatch: Record<string, string[]> =
     {};
 
-  constructor() {
-    const allowedParents = ['ATOMIC-RESULT-LIST', 'ATOMIC-FOLDED-RESULT-LIST'];
-    const isParentResultList = allowedParents.includes(
-      this.host.parentElement?.nodeName || ''
-    );
-
-    if (!isParentResultList) {
-      this.error = new Error(
-        'The "atomic-result-template" component has to be the child of either an "atomic-result-list" or an "atomic-folded-result-list" component.'
-      );
-      return;
-    }
-
-    if (!this.host.querySelector('template')) {
-      this.error = new Error(
-        'The "atomic-result-template" component has to contain a "template" element as a child.'
-      );
-    }
-
-    if (this.host.querySelector('template')?.content.querySelector('script')) {
-      console.warn(
-        'Any "script" tags defined inside of "template" elements are not supported and will not be executed when the results are rendered',
-        this.host
-      );
-    }
-  }
+  public resultTemplateCommon: ResultTemplateCommon;
 
   public componentWillLoad() {
-    for (const field in this.mustMatch) {
-      this.matchConditions.push(
-        ResultTemplatesHelpers.fieldMustMatch(field, this.mustMatch[field])
-      );
-    }
+    this.resultTemplateCommon.addMatchConditions(
+      this.mustMatch,
+      this.mustNotMatch
+    );
+  }
 
-    for (const field in this.mustNotMatch) {
-      this.matchConditions.push(
-        ResultTemplatesHelpers.fieldMustNotMatch(
-          field,
-          this.mustNotMatch[field]
-        )
-      );
-    }
+  constructor() {
+    this.resultTemplateCommon = new ResultTemplateCommon({
+      host: this.host,
+      setError: (err) => {
+        this.error = err;
+      },
+      validParents: ['atomic-result-list', 'atomic-folded-result-list'],
+      allowEmpty: true,
+    });
   }
 
   /**
    * Gets the appropriate result template based on conditions applied.
    */
   @Method()
-  public async getTemplate(): Promise<ResultTemplate<TemplateContent> | null> {
-    if (this.error) {
-      return null;
-    }
-
-    return {
-      conditions: this.getConditions(),
-      content: this.getContent(),
-      priority: 1,
-    };
-  }
-
-  private getConditions() {
-    return this.conditions.concat(this.matchConditions);
-  }
-
-  private getTemplateElement() {
-    return (
-      this.host.querySelector('template') ?? document.createElement('template')
-    );
-  }
-
-  private getContent() {
-    return this.getTemplateElement().content;
+  public async getTemplate(): Promise<ResultTemplate<DocumentFragment> | null> {
+    return this.resultTemplateCommon.getTemplate(this.conditions, this.error);
   }
 
   public render() {
-    if (this.error) {
-      return (
-        <atomic-component-error
-          element={this.host}
-          error={this.error}
-        ></atomic-component-error>
-      );
-    }
+    return this.resultTemplateCommon.renderIfError(this.error);
   }
 }