+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.13.0](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.12.11...@coveo/quantic@1.13.0) (2022-04-07)
+
+
+### Features
+
+* **headless, quantic:** search hub added to payload of case assist analytics" ([#1884](https://github.com/coveo/ui-kit/issues/1884)) ([a135b71](https://github.com/coveo/ui-kit/commit/a135b71c5e11ea25a351b0aeda8fb8c7b586d871))
+
+
+
+
+
 ## [1.12.11](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.12.10...@coveo/quantic@1.12.11) (2022-04-04)
 
 **Note:** Version bump only for package @coveo/quantic